# tests/bench_timers: A comprehensive benchmark for periph_timer