Depends on #8733 
(Excerpts from the [README](https://github.com/gebart/RIOT/blob/pr/periph-timer-stats/tests/bench_periph_timer/README.md))
# Benchmark test for periph_timer

This test is intended to collect statistics about the runtime delays in the
periph_timer implementation. This tool is mainly intended to detect problems in
the low level driver implementation which can be difficult to detect from
higher level systems such as xtimer.

## Test outline

The basic idea of the test is to generate a wide variety of calls to the
periph/timer API, in order to detect problems with the software implementation.
The test consists of a single thread of execution which will set random
timeouts on a periph_timer device. A reference timer is used to measure the
time it takes until the callback is called. The difference between the expected
time and the measured time is computed, and the mean and variance of the
recorded values are calculated. The results are printed as a table on stdout
every 30 seconds. All of the test scenarios used in this application are
based on experience from real world bugs encountered during the development of
RIOT and other systems.

### API functions tested

Both timer_set and timer_set_absolute calls are mixed in a random order, to
ensure that both functions are working correctly.

### Rescheduling

Some bugs may occur only when the application attempts to replace an already
set timer by calling timer_set again while the timer is running. This is dubbed
"resched" in this application. The application will issue rescheduling timer
calls in the normal operation of the test to catch this class of problems.

### Setting stopped timers

Another class of bugs is the kind where a timer is not behaving correctly if it
was stopped before setting a timer target. The timer should set the new target,
but not begin counting towards the target until timer_start is called. This is
covered by this application under the "stopped" category of test inputs.

### Avoiding phase lock

The CPU time used for generation and setting of the next timer target in
software is approximately constant across iterations. When the application flow
is driven by the timer under test, via a mutex which is unlocked from the timer
callback, the application will be "phase locked" to the timer under test. This
peculiarity may hide certain implementation race conditions which only occur at
specific phases of the timer ticks. In the test application, this problem is
avoided by inserting random CPU delay loops at key locations:

 - After timer_stop
 - Before timer_set/timer_set_absolute
 - Before timer_start

### Estimating benchmark CPU overhead

An estimation of the overhead resulting from the CPU processing delays inside
the benchmarking code is performed during benchmark initialization. The
estimation algorithm attempts to perform the exact same actions that the real
benchmark will perform, but without setting any timers. The measured delay is
assumed to originate in the benchmark code and will be subtracted when
computing the difference between expected and actual values. A warning is
printed when the variance of the estimated CPU overhead is too high, this can
be a sign that some other process is running on the CPU and disrupting the
estimation, or a sign of serious problems inside the timer_read function.

## Results

When the test has run for a certain amount of time, the current results will be
presented as a table on stdout. To assist with finding the source of any
discrepancies, the results are split according to three parameters:

 - Function used: timer_set, timer_set_absolute
 - Rescheduled: yes/no
 - Stopped: yes/no

### Interpreting timer_read statistics

A separate table is displayed for statistics on timer_read. This table compares
the expected timer under test time after a timer has triggered, against the
actual reported value from `timer_read(TIM_TEST_DEV)`. A positive value means
that the TUT time has passed the timer target time. A negative value means
that, according to the timer under test, the alarm target time has not yet been
reached, even though the timer callback has been executed.

### Example output

Below is a short sample of a test of a 32768 Hz timer with a 1 MHz reference:

    ------------- BEGIN STATISTICS --------------
    Limits: mean: [-10, 41], variance: [58, 99]
    Target error (actual trigger time - expected trigger time), in reference timer ticks
    positive: timer is late, negative: timer is early
    === timer_set running ===
       interval    count       sum       sum_sq    min   max  mean  variance
       1 -    2:   25705    455646      2002373      2    35    17     77
       3 -    4:   25533    457326      1973191      2    35    17     77
       5 -    8:   25412    451046      1970014      2    35    17     77
       9 -   16:   25699    458563      2017198      2    36    17     78
      17 -   32:   25832    457768      2018909      1    35    17     78
      33 -   64:   25758    440454      1978830      1    35    17     76
      65 -  128:   25335    414320      1942196      0    34    16     76
     129 -  256:   25254    367689      1979616     -3    32    14     78
     257 -  512:   25677    285822      2020736     -7    31    11     78
          TOTAL   230205   3788634     18459378     -7    36    16     80
...

The statistics above show that the timer implementation introduces a small
delay on short timers. Note however that it is not clear whether this delay is
in timer_set, or timer_read, but the combined effect of all error sources make
the timer overshoot its target by on average 17 microseconds. The difference
between min and max is close to the expected difference of approximately
1/32768 s = 30.51 µs. The lower min and mean for the longer timer intervals are
most likely caused by a drift between the 1 MHz clock and the 32.768 kHz clocks
inside the CPU, and is expected when using two separate clock sources for the
timers being compared.

Below is a short sample of a test where there is something wrong with the timer
implementation, again a 32768 Hz timer tested with a 1 MHz reference:

    === timer_set_absolute resched ===
       interval    count       sum       sum_sq    min   max  mean  variance
       1 -    2:  152217  82324291  13339518497     38  1071   540  87635  <=== SIC!
       3 -    4:  152199  82254909  13301794832     38  1071   540  87397  <=== SIC!
       5 -    8:  152023  82085454  13264873203     38  1071   539  87256  <=== SIC!
       9 -   16:  152156  82236539  13265351119     38  1071   540  87183  <=== SIC!
      17 -   32:  152639  82606815  13267666812     38  1071   541  86922  <=== SIC!
      33 -   64:  151883  81836155  13268661551     37  1070   538  87361  <=== SIC!
      65 -  128:  152114  82151495  13228100922     36  1070   540  86962  <=== SIC!
     129 -  256:  152363  81806042  13278055359     34  1069   536  87148  <=== SIC!
     257 -  512:  152360  81235185  13303811951     29  1066   533  87318  <=== SIC!
          TOTAL  1369954 738536885 119197349719     29  1071   539  87008  <=== SIC!

We can see that the variance, the maximum error, and the mean are very large.
This particular timer implementation needs some work on its timer_set_absolute
implementation.