@gebart 2018-03-22
This is ready for review now.
The simple test it started out as has grown to a very powerful benchmark tool for testing the periph_timer implementation and has been invaluable in verifying the performance of the old and new timer implementation for Kinetis LPTMR in #8814