@gebart 2018-04-11
Yes, it depends on the matstat module in #8733. I have updated the top post to reflect this.