@smlng 2018-05-30
on macOS the benchmark seems to hang, which isn't that surprising bc I regularly have problems with timers on RIOT native and macOS. Output so far, but nothing happening after long wait:

```
/RIOT/tests/bench_timers/bin/native/tests_bench_timers.elf  
RIOT native interrupts/signals initialized.
LED_RED_OFF
LED_GREEN_ON
RIOT native board initialized.
RIOT native hardware initialization complete.

main(): This is RIOT! (Version: 2018.07-devel-410-g857ea-mobi25.inet.haw-hamburg.de-HEAD)

Statistical benchmark for timers
Running timer test with seed 123 using Tiny Mersenne Twister PRNG.
TEST_MIN = 16
TEST_MAX = 128
TEST_MIN_REL = 0
TEST_MAX_REL = 112
TEST_NUM = 113
log2(TEST_NUM - 1) = 6
state vector elements per variant = 7
number of variants = 8
sizeof(state) = 32 bytes
state vector total memory usage = 1792 bytes
TIM_TEST_DEV = 0, TIM_TEST_FREQ = 1000000, TIM_TEST_CHAN = 0
TIM_REF_DEV  = 0, TIM_REF_FREQ  = 1000000
USE_REFERENCE = 1
TEST_PRINT_INTERVAL_TICKS = 30000000
Expected error variance due to truncation in tick conversion: 0
Calibrating spin delay...
spin_max = 12035
Estimating benchmark overhead...
overhead_target = 9 (s2 = 2)
overhead_read = 9 (s2 = 1)
exp_mask = 00000007
max interval = 255
```

no table with results, or is this expected?